Move some utils -> bytes.utils
Move functionality in `dask.utils` that was only/mainly used by bytes to
`bytes.utils`
